+++ NEWS.md
@@ -1,3 +1,5 @@
+# pxmake (development version)
+
 # pxmake 0.8.0
 
 ## Breaking changes