+++ NEWS.md
@@ -1,3 +1,5 @@
+# pxmake (development version)
+
 # pxmake 0.8.1
 
 ## Bug fixes and minor improvements